Google Classroom is a free web service developed by Gourob Dey
A classroom is a learning space in which both children and adults learn. Classrooms are found in educational institutions of all kinds, ranging from preschools to universities, and may also be found in other places where education or training is provided, such as corporations and religious and humanitarian organizations. The classroom provides a space where learning can take place uninterrupted by outside distractions.